Django/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
python-decouple==3.3
markdown==3.4.1
django-filter==22.1
requests
pandas
numpy